Improved print messages

+++ gcloudstoragemove/src/main/java/example/CloudStorageMove.java
@@ -1,11 +1,11 @@
 /**
- * CloudStorageMove is an example that handles Cloud Storage buckets on GCP (Google Cloud Platform)
+ * CloudStorageMove is an example that handles Cloud Storage buckets on GCP (Google Cloud Platform).
  * Move an object from a Cloud Storage bucket to another Cloud Storage bucket in a Google Cloud Project.
  * The application uses Application Default Credentials through a JSON service account key for authenticating.
  * The credentials are taken from GOOGLE_APPLICATION_CREDENTIALS environment variable.
- * You must use 3 parameters:
+ * You must provide 3 parameters:
  * SOURCE_BUCKET      = Source bucket name
- * SOURCE_FILE        = Source file name
+ * SOURCE_OBJECT      = Source object name
  * DESTINATION_BUCKET = Destination bucket name
  */
 
@@ -15,6 +15,7 @@ import java.io.IOException;
 
 import com.google.cloud.storage.Storage;
 import com.google.cloud.storage.StorageOptions;
+import com.google.cloud.storage.Bucket;
 import com.google.cloud.storage.Blob;
 import com.google.cloud.storage.Blob.BlobSourceOption;
 import com.google.cloud.storage.BlobId;
@@ -28,7 +29,7 @@ public class CloudStorageMove {
         String destinationBlobName;    // Destination blob name
 
         if (args.length < 3) {
-            System.out.println("Not enough parameters. Proper Usage is: java -jar gcloudstoragemove.jar <SOURCE_BUCKET> <SOURCE_FILE> <DESTINATION_BUCKET>");
+            System.out.println("Not enough parameters.\nProper Usage is: java -jar gcloudstoragemove.jar <SOURCE_BUCKET> <SOURCE_OBJECT> <DESTINATION_BUCKET>");
             System.exit(1);
         }
 
@@ -38,30 +39,38 @@ public class CloudStorageMove {
         destinationBlobName = sourceBlobName;
 
         System.out.println("From - bucket: " + sourceBucketName);
-        System.out.println("From - file:   " + sourceBlobName);
+        System.out.println("From - object: " + sourceBlobName);
         System.out.println("To   - bucket: " + destinationBucketName);
-        System.out.println("To   - file:   " + destinationBlobName);
+        System.out.println("To   - object: " + destinationBlobName);
 
         // Instantiates a client
         Storage storage = StorageOptions.getDefaultInstance().getService();
 
         // Copy a blob from a bucket to another bucket
-        System.out.println("Moving object...");
+        System.out.println("Moving object ...");
         Blob sourceBlob = storage.get(sourceBucketName, sourceBlobName);
 
-        CopyWriter copyWriter = sourceBlob.copyTo(BlobId.of(destinationBucketName, destinationBlobName));
-        Blob copiedBlob = copyWriter.getResult();
-
-        if (copiedBlob != null) {
-            // Delete blob from a source bucket
-            boolean deleted = sourceBlob.delete(BlobSourceOption.generationMatch());
-            if (deleted) {
-                System.out.println("Moved");
+        if (sourceBlob != null) {    // Exists source bucket/object
+            Bucket destinationBucket = storage.get(destinationBucketName);
+            if (destinationBucket != null) {   // Exists destination bucket
+                CopyWriter copyWriter = sourceBlob.copyTo(BlobId.of(destinationBucketName, destinationBlobName));
+                Blob copiedBlob = copyWriter.getResult();
+                if (copiedBlob != null) {
+                    // Delete blob from a source bucket
+                    boolean deleted = sourceBlob.delete(BlobSourceOption.generationMatch());
+                    if (deleted) {
+                        System.out.println("Moved");
+                    } else {
+                        System.out.println("Error: Object does NOT moved (not deleted)");
+                    }
+                } else {
+                    System.out.println("Error: Object does NOT moved (not copied)");
+                }
             } else {
-                System.out.println("Error: NOT moved (not deleted)");
+                System.out.println("Error: Destination Bucket does NOT exist.");
             }
         } else {
-            System.out.println("Error: NOT moved (not copied)");
+            System.out.println("Error: Source Bucket/Object does NOT exist.");
         }
     }
 }
